Enhance prompt selection with sequential logic
Added sequential selection logic and updated return values.
If sequential selection is used the logic checks the last used prompt or a value in the prompt name field.  - is there is none it starts at the first node in the json file and remembers the node. 
each run it will select the next prompt. if the end of the list is reached it will begin at the first prompt again. 
Also added a prompt_name output so the prompt name can be used as input for the save image prefix.